+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/domain/investigation/service/InvestigationServiceImpl.java
@@ -54,7 +54,6 @@ public class InvestigationServiceImpl extends AbstractQualityNotificationService
     @Override
     public QualityNotificationId start(StartQualityNotification startQualityAlertDomain) {
         QualityNotification notification = getNotificationPublisherService().startInvestigation(startQualityAlertDomain.getPartIds(), startQualityAlertDomain.getDescription(), startQualityAlertDomain.getTargetDate(), startQualityAlertDomain.getSeverity(), startQualityAlertDomain.getReceiverBpn(), startQualityAlertDomain.isAsBuilt());
-
         QualityNotificationId createdInvestigationId = getQualityNotificationRepository().saveQualityNotificationEntity(notification);
         log.info("Start Investigation {}", notification);
         return createdInvestigationId;