+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/qualitynotification/investigation/PublisherInvestigationsControllerIT.java
@@ -385,6 +385,7 @@ class PublisherInvestigationsControllerIT extends IntegrationTestSpecification {
                 .isAsBuilt(true)
                 .build();
 
+
         // when
         val investigationId = given()
                 .contentType(ContentType.JSON)