+++ CHANGELOG.md
@@ -11,8 +11,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ### Added
 - #515 Service Unavailable Response on Notification failure
-
 - #420 add /contracts api to fetch contract agreement information from EDC for assets
+- #423 Notification will be created/persisted in case of a edc notification
 
 ### Changed
 - Updated RELEASE.md to the latest release guide (added more steps)